backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
opencv-python-headless>=4.8.0,<4.10.0
numpy>=1.24.0,<2.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aiofiles>=23.2.1